moved file and added notes
