+++ src/main/java/frc/robot/Robot.java
@@ -5,6 +5,7 @@
 package frc.robot;
 
 import edu.wpi.first.wpilibj.TimedRobot;
+import frc.robot.Sensors.ColorSensor;
 
 /**
  * The VM is configured to automatically run this class, and to call the functions corresponding to
